CAMEL-23403: Move route diagram renderer to components/camel-diagram
Move RouteDiagramLayoutEngine and RouteDiagramRenderer from
camel-jbang-core to a new standalone components/camel-diagram module
so it can be reused by maven-plugin, Quarkus dev console, runtime
Camel apps, and 3rd-party tools.

Key changes:
- New module components/camel-diagram with public API
- printTextDiagram returns List<String> instead of using Printer
- RouteDiagramHelper utility for parsing route structure JSON
- camel-jbang-core depends on camel-diagram
