+++ bom/camel-bom/pom.xml
@@ -749,6 +749,11 @@
       </dependency>
       <dependency>
         <groupId>org.apache.camel</groupId>
+        <artifactId>camel-diagram</artifactId>
+        <version>4.21.0-SNAPSHOT</version>
+      </dependency>
+      <dependency>
+        <groupId>org.apache.camel</groupId>
         <artifactId>camel-digitalocean</artifactId>
         <version>4.21.0-SNAPSHOT</version>
       </dependency>
