+++ catalog/camel-allcomponents/pom.xml
@@ -639,6 +639,11 @@
         </dependency>
         <dependency>
             <groupId>org.apache.camel</groupId>
+            <artifactId>camel-diagram</artifactId>
+            <version>${project.version}</version>
+        </dependency>
+        <dependency>
+            <groupId>org.apache.camel</groupId>
             <artifactId>camel-digitalocean</artifactId>
             <version>${project.version}</version>
         </dependency>
