+++ components/pom.xml
@@ -120,6 +120,7 @@
         <module>camel-dapr</module>
         <module>camel-datasonnet</module>
         <module>camel-debug</module>
+        <module>camel-diagram</module>
         <module>camel-debezium</module>
         <module>camel-dhis2</module>
         <module>camel-digitalocean</module>
