+++ docs/components/modules/others/examples/json/diagram.json
@@ -1,0 +1,1 @@
+../../../../../../../../catalog/camel-catalog/src/generated/resources/org/apache/camel/catalog/others/diagram.json
