+++ dsl/camel-jbang/camel-jbang-core/src/main/java/org/apache/camel/dsl/jbang/core/commands/action/CamelRouteDiagramAction.java
@@ -26,20 +26,20 @@ import java.util.Optional;
 
 import javax.imageio.ImageIO;
 
+import org.apache.camel.diagram.RouteDiagramHelper;
+import org.apache.camel.diagram.RouteDiagramLayoutEngine;
+import org.apache.camel.diagram.RouteDiagramLayoutEngine.LayoutRoute;
+import org.apache.camel.diagram.RouteDiagramLayoutEngine.RouteInfo;
+import org.apache.camel.diagram.RouteDiagramRenderer;
+import org.apache.camel.diagram.RouteDiagramRenderer.DiagramColors;
 import org.apache.camel.dsl.jbang.core.commands.CamelJBangMain;
 import org.apache.camel.dsl.jbang.core.commands.Run;
-import org.apache.camel.dsl.jbang.core.commands.action.RouteDiagramLayoutEngine.LayoutRoute;
-import org.apache.camel.dsl.jbang.core.commands.action.RouteDiagramLayoutEngine.NodeInfo;
-import org.apache.camel.dsl.jbang.core.commands.action.RouteDiagramLayoutEngine.RouteInfo;
-import org.apache.camel.dsl.jbang.core.commands.action.RouteDiagramRenderer.DiagramColors;
 import org.apache.camel.dsl.jbang.core.common.CamelJBangConstants;
 import org.apache.camel.dsl.jbang.core.common.CommandLineHelper;
 import org.apache.camel.dsl.jbang.core.common.PathUtils;
 import org.apache.camel.main.KameletMain;
 import org.apache.camel.support.PatternHelper;
-import org.apache.camel.util.json.JsonArray;
 import org.apache.camel.util.json.JsonObject;
-import org.apache.camel.util.json.Jsoner;
 import org.jline.terminal.Terminal;
 import org.jline.terminal.TerminalBuilder;
 import org.jline.terminal.impl.TerminalGraphics;
@@ -213,12 +213,16 @@ public class CamelRouteDiagramAction extends ActionBaseCommand {
                                     "Terminal does not support graphics protocols (Kitty, iTerm2, or Sixel).");
                             printer().println(
                                     "Try running in a supported terminal: Kitty, iTerm2, WezTerm, Ghostty, or VS Code.");
-                            renderer.printTextDiagram(routes, printer());
+                            for (String line : renderer.printTextDiagram(routes)) {
+                                printer().println(line);
+                            }
                         }
                     } catch (IOException | UnsupportedOperationException e) {
                         printer().println("Failed to display diagram in terminal: " + e.getMessage());
                         printer().println("Falling back to text diagram.");
-                        renderer.printTextDiagram(routes, printer());
+                        for (String line : renderer.printTextDiagram(routes)) {
+                            printer().println(line);
+                        }
                     }
                 }
             }
@@ -275,31 +279,6 @@ public class CamelRouteDiagramAction extends ActionBaseCommand {
     }
 
     List<RouteInfo> parseRoutes(JsonObject jo) {
-        List<RouteInfo> routes = new ArrayList<>();
-        JsonArray arr = (JsonArray) jo.get("routes");
-        if (arr == null) {
-            return routes;
-        }
-
-        for (int i = 0; i < arr.size(); i++) {
-            JsonObject o = (JsonObject) arr.get(i);
-            RouteInfo route = new RouteInfo();
-            route.routeId = o.getString("routeId");
-            route.source = CamelRouteStructureAction.extractSourceName(o.getString("source"));
-
-            List<JsonObject> lines = o.getCollection("code");
-            if (lines != null) {
-                for (JsonObject line : lines) {
-                    NodeInfo node = new NodeInfo();
-                    node.type = line.getString("type");
-                    node.code = Jsoner.unescape(line.getString("code"));
-                    Integer level = line.getInteger("level");
-                    node.level = level != null ? level : 0;
-                    route.nodes.add(node);
-                }
-            }
-            routes.add(route);
-        }
-        return routes;
+        return RouteDiagramHelper.parseRoutes(jo);
     }
 }
